refactor(signing): extract party-key signing behind a trait
Introduce a `TransactionSigner` trait and a `signing` module so the
decentralized party's Daml-key signing has one clean seam that any
backend can plug into. The only operation that varies by backend is
turning an interactive-submission prepared hash into a Canton Signature;
everything else in the contracts sign step (loading the key, loading
prepared submissions, persisting the signature bundle) is
provider-independent and stays put.

This change is behaviour-preserving. The existing export-and-sign-locally
logic moves verbatim into `VaultExportSigner`, and `select_signer`
returns it unconditionally, so JCE parties sign exactly as before. The
signer is built from `NodeConfig::admin_channel` (same TLS-applying path
the step already used), not a bare URL, so the connection semantics are
unchanged too.

It sets up the KMS work (#264): a non-exportable KMS key can't use the
export backend, so adding AWS KMS / MPCH support becomes "implement
TransactionSigner and add a branch to select_signer" — no other changes
to the signing flow. `SigningKeyContext` already carries the
`kms_key_id` a KMS backend needs.

Refs #264, #241
