+++ crates/decman/Cargo.toml
@@ -42,6 +42,7 @@ actix-cors = { workspace = true }
 actix-web = { workspace = true }
 aes-gcm = { workspace = true }
 anyhow = { workspace = true }
+async-trait = { workspace = true }
 base64 = { workspace = true }
 bytes = { workspace = true }
 chrono = { workspace = true }
